requirements.txt
pytest
hypothesis
numpy
mido
scipy
pipe21
isort
hypothesis
flake8
autopep8
autoflake
pytest
pytest-cov
pytest-asyncio
pipe21
unify
Pillow
pyvips
